backend/requirements.txt
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.1
psycopg2-binary==2.9.9
djangorestframework-simplejwt==5.3.0
Pillow==10.1.0
reportlab==4.0.7
requests==2.31.0
python-dotenv==1.0.0
cryptography==41.0.7
google-generativeai>=0.3.0
web3==6.11.3
poetry>=1.5.0

